[QT-548] Overhaul remoteflight state and status checks (#181)
When investigating sporadic scenario failures in the Vault pipelines we
discovered that sometimes the vault or consul node is not in a valid
prerequisite state for us to perform the action we want to perform.
There are various reasons for this including:

  * The `consul_start` waiter simply restarted the service and didn't wait
    for it to be in a `READY` state. This will happen when `retry_join` has
    completed.
  * `go-discover` could sometimes take a while to discover all of the
    nodes so we don't have a functional back-end cluster when we attempt
    to initialize vault.
  * Sometimes systemd decides to restart services during start up if they
    don't enter a `READY` state fast enough.
  * We relied on `systemctl` sub-commands that could be unreliable. E.g.
    sometimes it can return non-LSB compliant codes or return valid
    codes for services that are "activating" rather than active,
    enabled, and running.
  * We relied on systemd to be the source of truth about a service,
    rather than a combination of the systemd's properties and various
    node status, healthy, and cluster APIs.

We've relied on the Terraform graph to cooridanate our actions
so that we only attempt things when prior actions have completed. In
those actions we did the minimum required state validation to complete.
We still require Terraforms graph to coordinate these actions, but we
are more aggressive about verifying pre-action state and verifying
post-action state. The result is a large refactor of our `remoteflight`
state checkers. Not only do our actions wait for the cluster or node to
be in an expected state, they also check the state afterwards to ensure
it's exactly as expected. The number of endpoints and types of state we
gather has grown significantly, as has the number of state checks that
we can do to enforce behavior. A bonus is that we now will return prior
and post state in human readable diagnostics for troubleshooting a failure.
We've also done a lot of the prerequisite work for a new `vault_status`
resource that we'll eventually be able to inject into any part of the
scenario graph to assert and/or wait for a desired state.

* Update Go modules.
* Bump VERSION.
* flightcontrol: add the ability to "download" an HTTP GET request to `STDOUT`.
* flightcontrol: dd the ability to exit with the HTTP status code for.
  "download" sub-command.
* retry: don't swallow errors during retry attempts.
* retry: don't swallow errors on retry.
* test: port the vault k8s scenario to this repository.
* test: run the vault k8s scenario before merge.
* lint: remove deprecated values from golangci-lint.
* kubernetes: add `ListPods()` to client.
* kubernetes: add human readable `String()` to pods.
* transport: move `TransportType` from `plugin` to `transport`.
* istrings: move `remoteflight.Indent()` to new `istrings` package.
* file: improve error messages in `CopyFile()`.
* remoteflight: support installing `enos-flight-control` in the target
  home directory.
* remoteflight: make `InstallFlightControl()` indempotent on targets
  that don't have `openssl` but have `sha256sum`.
* remoteflight: add `TargetHomeDir()` to find the targets home
  directory.
* remoteflight: add `TargetProcessManager()` to find the targets process
  manager.
* remoteflight: normalize transport variable name.
* k8s: make `transport` public.
* systemd: refactor client to use "show" instead of "is-active" and "status".
* systemd: rename `GetLogs()` to `GetUnitJournal()`
* systemd: clarify errors on failure.
* consul: add `/v1/agent/host` information to state.
* consul: add `/v1/health/node/:node` information to state.
* consul: add `/v1/health/state/passing` information to state.
* consul: add `/v1/operator/raft/configuration` information to state.
* consul: add state check for systemd properties.
* consul: add state check for node healthy.
* consul: add state check for cluster leader.
* consul: add state check for min healthy nodes.
* consul: add state check for min voters.
* consul: install `enos-flight-control` in home dir.
* consul: use `retry` package in `WaitForState()`.
* consul_start: wait for consul.service to be active and running in
  systemd.
* consul_start: wait for consul node to be healthy.
* consul_start: wait for consul cluster to have a leader.
* consul_start: wait for consul cluster to have expected number of
  nodes.
* consul_start: output extended state information in error diagnostics.
* vault: add `/v1/sys/storage/raft/autopilot/configuration` to state.
* vault: add `/v1/sys/storage/raft/autopilot/state` to state.
* vault: add `/v1/sys/config/state/sanitized` to state.
* vault: add `/v1/sys/replication` to state.
* vault: add `/v1/sys/health` to state.
* vault: add `/v1/sys/ha-status` to state.
* vault: add `/v1/sys/host-info` to state.
* vault: add `/v1/sys/storage/raft/configuration` to state.
* vault: add `/v1/sys/seal-status` to state.
* vault: add `vault status` to state.
* vault: add state check for systemd properties.
* vault: add state check for being initialized.
* vault: add state check for being sealed.
* vault: add state check for being unsealed.
* vault: add state check for health status.
* vault: add state check for UI being enabled in config.
* vault: add state check for a matching listener in config.
* vault: add state check for HA having an active node.
* vault: add state check for min HA nodes.
* vault: add state check for min raft servers.
* vault: add state check for min raft voters.
* vault: add state check for raft having a leader.
* vault: add state check for min autopilot servers.
* vault: add state check for autopilot being healthy.
* vault: add state check for autopilot having a leader.
* vault: add state check for storage type.
* vault: add state check for seal type.
* vault: install `enos-flight-control` in home dir.
* vault: use `retry` package in `WaitForState()`.
* vault_start: use `systemd_unit_name` when waiting for vault service
  state.
* vault_start: rely on status, seal-status, and health to determine the
  seal status.
* vault_start: rely on systemd unit properties instead of "is-active".
* vault_init: add `systemd_unit_name` to schema as it is required to
  verify the status before initializing.
* vault_init: rely on systemd unit properties instead of "is-active".
* vault_init: add before and after cluster state to diagnostics.
* vault_init: verify that cluster can be initialized before initializing.
* vault_init: rely on status and health to determine the init status.
* vault_init: allow waiting for k8s pod phases.
* vault_unseal: add `systemd_unit_name` to schema as it is required to
  verify the status before unsealing.
* vault_unseal: rely on systemd unit properties instead of "is-active".
* vault_unseal: add before and after cluster state to diagnostics.
* vault_unseal: verify that cluster can be unsealed before unsealing.
* vault_unseal: rely on status, seal-status, and health to determine the
  seal status.
* vault_unseal: allow waiting for k8s pod phases.
* vault_unseal: only assert seal type for shamir due to VAULT-7061

+++ internal/flightcontrol/download/command.go
@@ -2,6 +2,7 @@ package download
 
 import (
 	"context"
+	"errors"
 	"flag"
 	"fmt"
 	"io/fs"
@@ -28,15 +29,17 @@ func NewCommand(ui cli.Ui) (*Command, error) {
 
 // CommandArgs are the download commands arguments.
 type CommandArgs struct {
-	flags        *flag.FlagSet
-	destination  string
-	url          string
-	timeout      time.Duration
-	mode         int
-	sha256       string
-	authUser     string
-	authPassword string
-	replace      bool
+	flags                     *flag.FlagSet
+	destination               string
+	writeStdout               bool
+	url                       string
+	timeout                   time.Duration
+	mode                      int
+	sha256                    string
+	authUser                  string
+	authPassword              string
+	replace                   bool
+	exitWithRequestStatusCode bool
 }
 
 // Synopsis is the cli.Command synopsis.
@@ -53,14 +56,16 @@ Usage: enos-flight-control download --url https://some/remote/file.txt --destina
 
 Options:
 
-  --url            The URL of the file you wish to download
-  --destination    The local destination where you wish to write the file
-  --mode           The desired file permissions of the downloaded file
-  --timeout        The maximum allowable request time, eg: 1m
-  --sha256         Verifies that the downloaded file matches the given SHA 256 sum
-  --auth-user      The username to use for basic auth
-  --auth-password  The password to use for basic auth
-  --replace        Replace the destination file if it exists
+  --url                     The URL of the file you wish to download
+  --destination             The local destination where you wish to write the file
+  --stdout                  Write the output to stdout
+  --mode                    The desired file permissions of the downloaded file
+  --timeout                 The maximum allowable request time, eg: 1m
+  --sha256                  Verifies that the downloaded file matches the given SHA 256 sum
+  --auth-user               The username to use for basic auth
+  --auth-password           The password to use for basic auth
+  --replace                 Replace the destination file if it exists
+  --exit-with-status-code   On failure, exit with the HTTP status code returned
 
 `
 
@@ -76,18 +81,28 @@ func (c *Command) Run(args []string) int {
 	}
 
 	err = c.Download()
-	if err != nil {
-		c.ui.Error(err.Error())
-		return 1
+	if err == nil {
+		return 0
 	}
 
-	return 0
+	exitCode := 1
+	if c.args.exitWithRequestStatusCode {
+		var errDownload *ErrDownload
+		if errors.As(err, &errDownload) {
+			exitCode = errDownload.StatusCode
+		}
+	}
+
+	c.ui.Error(err.Error())
+
+	return exitCode
 }
 
 // Parse parses the raw args and maps them to the CommandArgs.
 func (a *CommandArgs) Parse(args []string) error {
 	a.flags = flag.NewFlagSet("download", flag.ContinueOnError)
 	a.flags.StringVar(&a.destination, "destination", "", "where to write the resulting file")
+	a.flags.BoolVar(&a.writeStdout, "stdout", false, "Write the output to stdout")
 	a.flags.StringVar(&a.url, "url", "", "the URL of the resource you wish to download")
 	a.flags.IntVar(&a.mode, "mode", 0o666, "the file permissions of the downloaded file")
 	a.flags.DurationVar(&a.timeout, "timeout", 5*time.Minute, "the maximum allowable request time")
@@ -95,10 +110,15 @@ func (a *CommandArgs) Parse(args []string) error {
 	a.flags.StringVar(&a.authUser, "auth-user", "", "if given, sets the basic auth username when making the HTTP request")
 	a.flags.StringVar(&a.authPassword, "auth-password", "", "if given, sets the basic auth password when making the HTTP request")
 	a.flags.BoolVar(&a.replace, "replace", false, "overwite the destination if it already exists")
+	a.flags.BoolVar(&a.exitWithRequestStatusCode, "exit-with-status-code", false, "On failure, exit with the HTTP status code returned")
 
 	err := a.flags.Parse(args)
 	if err != nil {
 		return err
+	}
+
+	if !a.writeStdout && a.destination == "" {
+		return fmt.Errorf("you must provide either a destination or stdout")
 	}
 
 	return nil
@@ -109,29 +129,33 @@ func (c *Command) Download() error {
 	ctx, cancel := context.WithTimeout(context.Background(), c.args.timeout)
 	defer cancel()
 
-	_, err := os.Stat(c.args.destination)
-	if err == nil {
-		// The destination file already exists
-		if !c.args.replace {
-			return fmt.Errorf("%s already exists. Set --replace=true to replace existing files", c.args.destination)
+	opts := []RequestOpt{
+		WithRequestURL(c.args.url),
+		WithRequestSHA256(c.args.sha256),
+		WithRequestWriteStdout(c.args.writeStdout),
+	}
+
+	if c.args.destination != "" {
+		_, err := os.Stat(c.args.destination)
+		if err == nil {
+			// The destination file already exists
+			if !c.args.replace {
+				return fmt.Errorf("%s already exists. Set --replace=true to replace existing files", c.args.destination)
+			}
+
+			err = os.Remove(c.args.destination)
+			if err != nil {
+				return err
+			}
 		}
 
-		err = os.Remove(c.args.destination)
+		dst, err := os.OpenFile(c.args.destination, os.O_RDWR|os.O_CREATE, fs.FileMode(c.args.mode))
 		if err != nil {
 			return err
 		}
-	}
+		defer dst.Close()
 
-	dst, err := os.OpenFile(c.args.destination, os.O_RDWR|os.O_CREATE, fs.FileMode(c.args.mode))
-	if err != nil {
-		return err
-	}
-	defer dst.Close()
-
-	opts := []RequestOpt{
-		WithRequestURL(c.args.url),
-		WithRequestDestination(dst),
-		WithRequestSHA256(c.args.sha256),
+		opts = append(opts, WithRequestDestination(dst))
 	}
 
 	if c.args.authUser != "" {
